+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/components/ConfirmReceiptStatus.tsx
@@ -1,5 +1,3 @@
-"use client";
-import React from "react";
 import { useTranslation } from "@i18n/client";
 import { ExclamationText } from "@clientComponents/form-builder/app/shared";
 import { VaultStatus } from "@lib/types";
