+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/components/RemovalStatus.tsx
@@ -1,5 +1,3 @@
-"use client";
-import React from "react";
 import { useTranslation } from "@i18n/client";
 import { VaultStatus } from "@lib/types";
 import { getDaysPassed } from "@lib/client/clientHelpers";
