+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/components/Responses.tsx
@@ -1,24 +1,15 @@
 "use client";
-import { useState } from "react";
-import { useTranslation } from "@i18n/client";
-import { FormRecord, VaultSubmissionList } from "@lib/types";
-import { useSession } from "next-auth/react";
-import Link from "next/link";
-
-import { DownloadTable } from "./DownloadTable";
-import { ReportDialog } from "./Dialogs/ReportDialog";
-import { NagwareResult } from "@lib/types";
 import { Nagware } from "@clientComponents/form-builder/app/Nagware";
 import { ClosedBanner } from "@clientComponents/form-builder/app/shared/ClosedBanner";
-import { WarningIcon } from "@serverComponents/icons";
-import { useRouter, useParams, usePathname } from "next/navigation";
-
-import { ConfirmDialog } from "./Dialogs/ConfirmDialog";
-import { Alert } from "@clientComponents/globals";
-
 import { useRehydrate } from "@clientComponents/form-builder/hooks";
+import { useTranslation } from "@i18n/client";
 import { ucfirst } from "@lib/client/clientHelpers";
-import { TitleAndDescription } from "./TitleAndDescription";
+import { FormRecord, NagwareResult, VaultSubmissionList } from "@lib/types";
+import { useSession } from "next-auth/react";
+import Link from "next/link";
+import { useParams } from "next/navigation";
+import { ReportDialog } from "./Dialogs/ReportDialog";
+import { DownloadTable } from "./DownloadTable";
 import { NoResponses } from "./NoResponses";
 
 export interface ResponsesProps {
@@ -45,17 +36,10 @@ export const Responses = ({
   } = useTranslation("form-builder-responses");
   const { status } = useSession();
   const isAuthenticated = status === "authenticated";
-  const [isShowReportProblemsDialog, setIsShowReportProblemsDialog] = useState(false);
-  const [showConfirmReceiptDialog, setShowConfirmReceiptDialog] = useState(false);
-  const [successAlertMessage, setShowSuccessAlert] = useState<false | string>(false);
-  const router = useRouter();
-
   const { id, statusFilter: rawStatusFilter } = useParams<{ id: string; statusFilter: string }>();
   const statusFilter = ucfirst(rawStatusFilter);
 
   const formId = id;
-
-  const pathName = usePathname();
 
   let formName = "";
   if (initialForm) {
@@ -71,20 +55,6 @@ export const Responses = ({
 
   return (
     <>
-      {isAuthenticated && vaultSubmissions.length > 0 && (
-        <TitleAndDescription
-          setShowConfirmReceiptDialog={setShowConfirmReceiptDialog}
-          statusFilter={statusFilter}
-        />
-      )}
-
-      {successAlertMessage && (
-        <Alert.Success className="mb-4">
-          <Alert.Title>{t(`${successAlertMessage}.title`)}</Alert.Title>
-          <Alert.Body>{t(`${successAlertMessage}.body`)}</Alert.Body>
-        </Alert.Success>
-      )}
-
       {nagwareResult && <Nagware nagwareResult={nagwareResult} />}
 
       {isAuthenticated && (
@@ -99,8 +69,6 @@ export const Responses = ({
                   formId={formId}
                   nagwareResult={nagwareResult}
                   responseDownloadLimit={responseDownloadLimit}
-                  showDownloadSuccess={successAlertMessage}
-                  setShowDownloadSuccess={setShowSuccessAlert}
                   lastEvaluatedKey={lastEvaluatedKey}
                 />
               </>
@@ -108,16 +76,10 @@ export const Responses = ({
             {vaultSubmissions.length <= 0 && <NoResponses statusFilter={statusFilter} />}
           </div>
           <div className="mt-8">
-            <Link
-              onClick={() => setIsShowReportProblemsDialog(true)}
-              href={"#"}
-              className="text-black visited:text-black focus:text-white-default"
-              id="reportProblemButton"
-            >
-              <WarningIcon className="mr-2 inline-block" />
-              {t("responses.reportProblems")}
-            </Link>
-
+            <ReportDialog
+              apiUrl={`/api/id/${formId}/submission/report`}
+              maxEntries={MAX_REPORT_COUNT}
+            />
             <Link
               href={`/form-builder/responses/${formId}/problem`}
               className="ml-12 text-black visited:text-black focus:text-white-default"
@@ -144,24 +106,6 @@ export const Responses = ({
           </p>
         </>
       )}
-
-      <ReportDialog
-        isShow={isShowReportProblemsDialog}
-        setIsShow={setIsShowReportProblemsDialog}
-        apiUrl={`/api/id/${formId}/submission/report`}
-        maxEntries={MAX_REPORT_COUNT}
-      />
-
-      <ConfirmDialog
-        isShow={showConfirmReceiptDialog}
-        setIsShow={setShowConfirmReceiptDialog}
-        apiUrl={`/api/id/${formId}/submission/confirm`}
-        maxEntries={responseDownloadLimit}
-        onSuccessfulConfirm={() => {
-          router.replace(pathName, { scroll: false });
-          setShowSuccessAlert("confirmSuccess");
-        }}
-      />
     </>
   );
 };
