+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/layout.tsx
@@ -1,7 +1,11 @@
+import { isEmailDelivery } from "@clientComponents/form-builder/util";
+import { getAppSetting } from "@lib/appSettings";
+import { auth } from "@lib/auth";
+import { ucfirst } from "@lib/client/clientHelpers";
 import { DeliveryOptionEmail } from "./components/DeliveryOptionEmail";
 import { NavigationTabs } from "./components/NavigationTabs";
-import { isEmailDelivery } from "@clientComponents/form-builder/util";
-import { fetchTemplate } from "./utils";
+import { TitleAndDescription } from "./components/TitleAndDescription";
+import { fetchSubmissions, fetchTemplate } from "./utils";
 
 export default async function Layout({
   children,
@@ -11,9 +15,16 @@ export default async function Layout({
   params: { locale: string; statusFilter: string; id: string };
 }) {
   const initialForm = await fetchTemplate(id);
-
+  const session = await auth();
+  const isAuthenticated = session !== null;
+  const responseDownloadLimit = Number(await getAppSetting("responseDownloadLimit"));
   const deliveryOption = initialForm?.deliveryOption;
   const isPublished = initialForm?.isPublished || false;
+
+  const { submissions } = await fetchSubmissions({
+    formId: id,
+    status: statusFilter,
+  });
 
   if (deliveryOption && isEmailDelivery(deliveryOption)) {
     return (
@@ -32,6 +43,13 @@ export default async function Layout({
   return (
     <>
       <NavigationTabs statusFilter={statusFilter} formId={id} locale={locale} />
+      {isAuthenticated && submissions.length > 0 && (
+        <TitleAndDescription
+          statusFilter={ucfirst(statusFilter)}
+          formId={id}
+          responseDownloadLimit={responseDownloadLimit}
+        />
+      )}
       {children}
     </>
   );
